Changed return value from long to int

+++ src/main/java/edu/hw1/Task1.java
@@ -8,7 +8,7 @@ import edu.Task;
 public final class Task1 extends Task {
     private Task1() {}
 
-    public static long minutesToSeconds(String str) {
+    public static int minutesToSeconds(String str) {
 //        Паттерн с допускание строки вида "00:00"
         final String regex = "^(?!0+:00)\\d{2,}:[0-5]\\d$";
         final Pattern pattern = Pattern.compile(regex);
@@ -18,7 +18,7 @@ public final class Task1 extends Task {
             return -1;
         }
         final String[] value = str.split(":");
-        final long minutes = Long.parseLong(value[0]) * 60;
+        final int minutes = Integer.parseInt(value[0]) * 60;
         final int seconds = Integer.parseInt(value[1]);
         return minutes + seconds;
     }
